ImageLLMAnalyzer/requirements.txt
openai
dotenv
questionary
einops
pypdf
logging
python-telegram-bot
